其他模型修复编译问题

+++ backend/src/main/java/com/handmade/strategy/WorkSimilarityCalculator.java
@@ -39,10 +39,10 @@ public class WorkSimilarityCalculator {
 
     public long calculate(Work baseWork, Work candidateWork) {
         if (baseWork == null || candidateWork == null) {
-            return MIN_SCORE;
+            return (long) MIN_SCORE;
         }
         if (baseWork.getId() != null && baseWork.getId().equals(candidateWork.getId())) {
-            return MIN_SCORE;
+            return (long) MIN_SCORE;
         }
 
         double categoryScore = calculateCategoryScore(baseWork, candidateWork);
